Refactor: changed the type of SafeKwhRead to a double to show fractional values rather than whole numbers. Also removed Serial printing in main.cpp

+++ lib/PulseSensor/PulseSensor.cpp
@@ -6,7 +6,7 @@ volatile unsigned long lastPulseTime = 0;
 volatile unsigned long rawFires = 0;
 volatile unsigned long totalPulses = 0;
 
-unsigned long safeKwhRead = 0;
+double safeKwhRead = 0;
 
 void initSensor() {
   pinMode(SENSOR_PIN, INPUT);
